Update lock scripts for new xidlehook.

+++ lock-timer.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/usr/bin/env bash
 
 if [ -z "$1" ]; then
     echo "No timer command supplied. Exiting..."
@@ -9,12 +9,13 @@ dir="$(dirname "$0")"
 
 pkill xidlehook
 
-rm /tmp/xidlehook.sock
+if [ -e /tmp/xidlehook.sock ]; then
+    rm /tmp/xidlehook.sock
+fi
+
+export PRIMARY_DISPLAY="$(xrandr | grep primary | cut -d ' ' -f 1)"
 
 read -r -d '' lock << EOF
-    # Undim screen
-    xrandr --output "\$(xrandr | grep primary | cut -d ' ' -f 1)" --brightness 1
-
     # Pause notifications
     pkill dunst -USR1
 
@@ -23,13 +24,21 @@ read -r -d '' lock << EOF
     # Start notifications
     pkill dunst -USR2
 EOF
+export lock
 
+    # --not-when-audio \
 xidlehook \
-  --time 15 \
-  --timer "$lock" \
-  --notify 10 \
-  --notifier  'xrandr --output "$(xrandr | grep primary | cut -d " " -f 1)" --brightness .1' \
-  --canceller 'xrandr --output "$(xrandr | grep primary | cut -d " " -f 1)" --brightness 1' \
-  --not-when-fullscreen \
-  --not-when-audio \
-  --socket /tmp/xidlehook.sock
+    --not-when-fullscreen \
+    --socket /tmp/xidlehook.sock \
+    --timer 600 \
+        'xrandr --output "$PRIMARY_DISPLAY" --brightness .1' \
+        'xrandr --output "$PRIMARY_DISPLAY" --brightness 1' \
+    --timer 10 \
+        'xrandr --output "$PRIMARY_DISPLAY" --brightness 1 && /bin/sh -c "$lock"' \
+        '' \
+    --timer 60 \
+        'xset dpms force off' \
+        '' \
+    --timer 600 \
+        'systemctl suspend' \
+        '' \
